Cleanup
Replaced (almost) all occurrences of strings with numeric literals
when setting limits or a fixed value.

Strings were originally used in an attempt to boost confidence in the
results of the tests, because that seemed to make the tests resemble
manual XML configuration more closely. But if there were truly a
difference, then why even support the Object arguments instead of
limiting arguments to Strings?